+++ fsic/core/linkers.py
@@ -192,16 +192,34 @@ Spans of submodels differ:
         )
 
     def to_dataframe(
-        self, *, status: bool = True, iterations: bool = True
+        self,
+        *,
+        status: bool = True,
+        iterations: bool = True,
+        include_internal: bool = False,
     ) -> 'pandas.DataFrame':  # noqa: F821
         """Return the values and solution information from the linker as a `pandas` DataFrame. **Requires `pandas`**."""
-        return _model_to_dataframe(self, status=status, iterations=iterations)
+        return _model_to_dataframe(
+            self,
+            status=status,
+            iterations=iterations,
+            include_internal=include_internal,
+        )
 
     def to_dataframes(
-        self, *, status: bool = True, iterations: bool = True
+        self,
+        *,
+        status: bool = True,
+        iterations: bool = True,
+        include_internal: bool = False,
     ) -> Dict[Hashable, 'pandas.DataFrame']:  # noqa: F821
         """Return the values and solution information from the linker and its constituent submodels as `pandas` DataFrames. **Requires `pandas`**."""
-        return _linker_to_dataframes(self, status=status, iterations=iterations)
+        return _linker_to_dataframes(
+            self,
+            status=status,
+            iterations=iterations,
+            include_internal=include_internal,
+        )
 
     def solve(
         self,