+++ CHANGELOG.md
@@ -27,6 +27,9 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
   `VectorContainer._locate_period_in_span()`, to then raise as `KeyError`s.
 - Added check (and error) in `iter_periods()` if attempting to solve an
   instance with an empty span i.e. if there are no periods to solve.
+- Added new keyword argument `include_internal` to control
+  `model_to_dataframe()` handling of internal model variables with names
+  prefixed with an underscore.
 - Added new `TracerMixin` extension for `BaseModel`-derived classes, to track
   iteration-by-iteration variable results.